Fix timestamp decoding for 32-bit architectures. (#528)
We were converting timestamps to NSInteger, which is a 32-bit or 64-bit
int depending on the architecture. `long long` is guaranteed to be
64-bit wide.

Fixes #525.

+++ Source/ARTNSDate+ARTUtil.h
@@ -11,7 +11,7 @@
 @interface NSDate (ARTUtil)
 
 + (instancetype)artDateFromNumberMs:(NSNumber *)number;
-+ (instancetype)artDateFromIntegerMs:(NSInteger)ms;
++ (instancetype)artDateFromIntegerMs:(long long)ms;
 
 - (NSNumber *)artToNumberMs;
 - (NSInteger)artToIntegerMs;
